requirements.txt
duckdb>=0.10.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0
yfinance>=0.2.36
requests>=2.31.0
pytrends>=4.9.2
PyYAML>=6.0.1
schedule>=1.2.0
streamlit>=1.32.0
pytest>=8.0.0
python-dotenv>=1.0.0
